refactor: stirling1.hpp を stirling.hpp にリネーム
第1種・第2種両方を含むファイルになったため、名前を中身に合わせた。
