+++ lib/combinatorics/stirling.hpp
