+++ test/yosupo/enumerative_conbinatorics/stirling_number_of_the_first_kind_fixed_k.test.cpp
@@ -1,6 +1,6 @@
 // competitive-verifier: PROBLEM https://judge.yosupo.jp/problem/stirling_number_of_the_first_kind_fixed_k
 #include <iostream>
-#include "combinatorics/stirling1.hpp"
+#include "combinatorics/stirling.hpp"
 
 using Mint = modint998;
